refacto: introduce thermal_modelisation.py file
